安全军规告警修复

+++ arkweb/ohos_nweb/src/cef_delegate/nweb_delegate.cc
@@ -2729,7 +2729,7 @@ void NWebDelegate::RegisterNativeScrollCallback(
 #if BUILDFLAG(ARKWEB_JSPROXY)
 void NWebDelegate::JavaScriptOnDocumentStart(const ScriptItems& scriptItems) {
   if (GetBrowser() != nullptr && GetBrowser()->GetHost() != nullptr) {
-    int count = 0;
+    size_t count = 0;
     if (scriptItems.size() == 0) {
       GetBrowser()->GetHost()->JavaScriptOnDocumentStart("", std::vector<CefString>(),
                                                          true);
@@ -2761,7 +2761,7 @@ void NWebDelegate::JavaScriptOnDocumentStartByOrder(
       GetBrowser()->GetHost()->JavaScriptOnDocumentStart("", std::vector<CefString>(),
                                                          true);
     }
-    int count = 0;
+    size_t count = 0;
     for (const auto& item : scriptItemsByOrder) {
       if (scriptItems.find(item) == scriptItems.end()) {
         continue;
@@ -2793,7 +2793,7 @@ void NWebDelegate::JavaScriptOnDocumentEndByOrder(
       GetBrowser()->GetHost()->JavaScriptOnDocumentEnd("", std::vector<CefString>(),
                                                        true);
     }
-    int count = 0;
+    size_t count = 0;
     for (const auto& item : scriptItemsByOrder) {
       if (scriptItems.find(item) == scriptItems.end()) {
         continue;
@@ -2825,7 +2825,7 @@ void NWebDelegate::JavaScriptOnHeadReadyByOrder(
       GetBrowser()->GetHost()->JavaScriptOnHeadReady("", std::vector<CefString>(),
                                                      true);
     }
-    int count = 0;
+    size_t count = 0;
     for (const auto& item : scriptItemsByOrder) {
       if (scriptItems.find(item) == scriptItems.end()) {
         continue;
@@ -2903,7 +2903,7 @@ void NWebDelegate::JavaScriptOnDocumentEnd(const ScriptItems& scriptItems) {
       GetBrowser()->GetHost()->JavaScriptOnDocumentEnd("", std::vector<CefString>(),
                                                        true);
     }
-    int count = 0;
+    size_t count = 0;
     for (auto item : scriptItems) {
       count++;
       CefString script = item.first;
